+++ src/Network/Network.Test/ScenarioTests/CortexTests.ps1
@@ -101,7 +101,7 @@ function Test-CortexCRUD
 		Assert-AreEqual $virtualHubName $virtualHub.Name
 		$routes = $virtualHub.RouteTable.Routes
 		Assert-AreEqual 2 @($routes).Count
-		
+
 		# Create the VpnSite
 		$vpnSiteAddressSpaces = New-Object string[] 1
 		$vpnSiteAddressSpaces[0] = "192.168.2.0/24"
@@ -851,6 +851,51 @@ function Test-CortexExpressRouteCRUD
      }
 }
 
+function Test-CortexVirtualHubCRUD
+{
+	# Setup
+  $rgName = Get-ResourceName
+  $rglocation = Get-ProviderLocation ResourceManagement "West Central US"
+
+	$virtualWanName = Get-ResourceName
+	$virtualHubName = Get-ResourceName
+
+	try
+	{
+		# Create the resource group
+    $resourceGroup = New-AzResourceGroup -Name $rgName -Location $rglocation
+
+		# Create the Virtual Wan
+		$createdVirtualWan = New-AzVirtualWan -ResourceGroupName $rgName -Name $virtualWanName -Location $rglocation -AllowVnetToVnetTraffic -AllowBranchToBranchTraffic
+		$virtualWan = Get-AzVirtualWan -ResourceGroupName $rgName -Name $virtualWanName
+		Assert-AreEqual $rgName $virtualWan.ResourceGroupName
+		Assert-AreEqual $virtualWanName $virtualWan.Name
+		Assert-AreEqual $true $virtualWan.AllowVnetToVnetTraffic
+		Assert-AreEqual $true $virtualWan.AllowBranchToBranchTraffic
+
+		# Create the Virtual Hub
+		$createdVirtualHub = New-AzVirtualHub -ResourceGroupName $rgName -Name $virtualHubName -Location $rglocation -AddressPrefix "10.0.0.0/16" -VirtualWan $virtualWan
+		$virtualHub = Get-AzVirtualHub -ResourceGroupName $rgName -Name $virtualHubName
+		Assert-AreEqual $rgName $virtualHub.ResourceGroupName
+		Assert-AreEqual $virtualHubName $virtualHub.Name
+		Assert-AreEqual "10.0.0.0/16" $virtualHub.AddressPrefix
+
+		# Reset-AzHubRouter
+		Reset-AzHubRouter -ResourceGroupName $rgName -Name $virtualHubName
+
+		# Delete the resources
+    $delete = Remove-AzVirtualHub -ResourceGroupName $rgName -Name $virtualHubName -Force -PassThru
+    Assert-AreEqual $True $delete
+
+    $delete = Remove-AzVirtualWan -ResourceGroupName $rgName -Name $virtualWanName -Force -PassThru
+    Assert-AreEqual $True $delete
+	}
+	finally
+	{
+		Clean-ResourceGroup $rgname
+	}
+}
+
 function Test-VHubRouteTableCRUD {
 	# Setup
 	$rgName = Get-ResourceName
@@ -904,7 +949,7 @@ function Test-VHubRouteTableCRUD {
 		Assert-AreEqual 1 $customRouteTable.Labels.Count
 
 		# Delete the custom route table
-		$ delete = Remove-AzVHubRouteTable -ResourceGroupName $rgName -VirtualHubName $virtualHubName -Name $customRouteTableName -Force -PassThru
+		$delete = Remove-AzVHubRouteTable -ResourceGroupName $rgName -VirtualHubName $virtualHubName -Name $customRouteTableName -Force -PassThru
 		Assert-AreEqual $True $delete
 	}
 	finally